MLIBZ-2316: Server Side Delta Set (#270)
* MLIBZ-2316: Delete deltaFetch.js

* MLIBZ-2316: Add deltaset.js

* MLIBZ-2316: Remove export of deltaFetch

* MLIBZ-2316: Use new deltaSet function in NetworkRepository

* Remove TODO comment

* MLIBZ-2316: Add clearDeltaSet function

* MLIBZ-2316: Add tests for using delta set with a CacheStore

* MLIBZ-2316: Rename useDeltaFetch to useDeltaSet

* MLIBZ-2316: Add delta set constants

* MLIBZ-2316: Split delta set logic into smaller functions

* MLIBZ-2316: Add tests for delta set

* Update package versions to 3.10.1-beta.1

* Make requested changes from code review for delta set

* Don't perform _replaceOfflineEntities() when using delta set

* Fix typo and change name from xKinveyRequestStartHeader to kinveyRequestStartHeader

* Don't duplicate function calls in SyncManager.pull()

* Fix check for if statment for SyncManager.pull()

* Only replace entities if useDeltaSet is not true

* Default options to an empty object in SyncManager.pull()

* Fix failing unit tests related to changes

* Add checks for skip amd limit modifiers to delta set

* MLIBZ-2316: Handle skip and limit for Delta Set

We will log that the skip and limit modifiers will be ignored for a delta set request. These modifiers will be appliead to the cache after the delta set request has completed. A find() and pull() on a data store will now honor skip and limit modifiers preventing the user from keeping track of when they can and cannot use a specific query.

* Update logged message for delta set

* Use original query when performing regular GET request for delta set

* MLIBZ-2316: Send a regular GET request if a query contains a skip and/or limit modifier

* MLIBZ-2437: Send a regular GET request if a ParamaterValueOutOfRangeError is received from a delta set request

* Add missing curly brace

* Fixing failing tests

* Add InvalidCachedQuery error

* Export InvalidCachedQuery error

* Add querycache.js

* Add a getter for X-Kinvey-Request-Start header to Headers class

* Remove constants

* Refactor SyncManager to use new deltaset and querycache files

* Use requestStart getter from headers

* Fix failing unit tests

* Bug fixes for Delta Set

* Fix some bugs found during testing

* Fix skipped unit tests

* MLIBZ-2478: Return count of items changed when making a _deltaset request with pull()

* MLIBZ-2477: Handle NotFoundError when deleting a cachedQuery

* MLIBZ-2486: Ignore limit and skip for auto pagination and update tests

* MLIBZ-2493: Perform delta set delete and then delta set changed serially instead of in parallel.

* MLIBZ-2449: Do not filter out _QueryCache table or __testSupport__ table when get all tables for WebSQL

* MLIBZ-2495: Ignore deleting entities when limit and skip are used to pull entities

* Update _queryCache after auto pagination has completed

* MLIBZ-2512: Ignore skip value of query for auto pagination

* Fix failing unit test that checks if skip is ignored for auto pagination

* Added deltaset tests (#294)

* Added deltaset tests

* added new tests for limit and skip and some unit tests

* skipped tests that use websql to not run on other than html5 and removed a .only

* removed another .only

* did some refactoring and fixed tests failing for Nativescript and Phonegap

* Fix status code set with nock for failing unit test

* Added deltaset tests

* added new tests for limit and skip and some unit tests

* skipped tests that use websql to not run on other than html5 and removed a .only

* removed another .only

* did some refactoring and fixed tests failing for Nativescript and Phonegap

* Fix status code set with nock for failing unit test

* added a test for autopagination with skip and limit

* fixing an expectation string and wrong expectation for deltaset test

* Added deltaset tests

* added new tests for limit and skip and some unit tests

* skipped tests that use websql to not run on other than html5 and removed a .only

* removed another .only

* did some refactoring and fixed tests failing for Nativescript and Phonegap

* Fix status code set with nock for failing unit test

* removed .only again

+++ src/core/datastore/sync/sync-manager.js
@@ -2,23 +2,16 @@ import { Promise } from 'es6-promise';
 import clone from 'lodash/clone';
 
 import { Log } from '../../log';
-import { KinveyError, NotFoundError, SyncError } from '../../errors';
-
+import { KinveyError, NotFoundError, SyncError, InvalidCachedQuery } from '../../errors';
 import { getPlatformConfig } from '../../platform-configs';
 import { SyncOperation } from './sync-operation';
 import { maxEntityLimit, defaultPullSortField } from './utils';
 import { isEmpty } from '../utils';
 import { repositoryProvider } from '../repositories';
 import { Query } from '../../query';
-import {
-  ensureArray,
-  isNonemptyString,
-  forEachAsync,
-  splitQueryIntoPages
-} from '../../utils';
-
-// imported for typings
-// import { SyncStateManager } from './sync-state-manager';
+import { ensureArray, isNonemptyString, forEachAsync, splitQueryIntoPages } from '../../utils';
+import { deltaSet } from '../deltaset';
+import { getCachedQuery, updateCachedQuery, deleteCachedQuery } from '../querycache';
 
 const {
   maxConcurrentPullRequests: maxConcurrentPulls,
@@ -67,18 +60,82 @@ export class SyncManager {
       });
   }
 
-  pull(collection, query, options) {
-    if (!isNonemptyString(collection)) {
-      return Promise.reject(new KinveyError('Invalid or missing collection name'));
-    }
+  pull(collection, query, options = {}) {
+    return Promise.resolve()
+      .then(() => {
+        if (!isNonemptyString(collection)) {
+          throw new KinveyError('Invalid or missing collection name');
+        }
+      })
+      .then(() => {
+        if (options.useDeltaSet) {
+          return deltaSet(collection, query, options)
+            .then((response) => {
+              return getCachedQuery(collection, query)
+                .then((cachedQuery) => {
+                  if (cachedQuery) {
+                    cachedQuery.lastRequest = response.headers.requestStart;
+                    return updateCachedQuery(cachedQuery);
+                  }
 
-    if (options && (options.autoPagination && !options.useDeltaFetch)) {
-      return this._paginatedPull(collection, query, options);
-    }
+                  return null;
+                })
+                .then(() => response.data);
+            })
+            .then((data) => {
+              if (data.deleted.length > 0) {
+                const deleteQuery = new Query();
+                deleteQuery.containsAll('_id', data.deleted.map((entity) => entity._id));
+                return this._deleteOfflineEntities(collection, deleteQuery)
+                  .then(() => data);
+              }
 
-    return this._fetchItemsFromServer(collection, query, options)
-      .then(entities => this._replaceOfflineEntities(collection, query, entities))
-      .then(replacedEntities => replacedEntities.length);
+              return data;
+            })
+            .then((data) => {
+              if (data.changed.length > 0) {
+                return this._getOfflineRepo()
+                  .then((offlineRepo) => offlineRepo.update(collection, data.changed))
+                  .then(() => data.changed.length);
+              }
+
+              return 0;
+            });
+        } else if (options.autoPagination) {
+          return this._paginatedPull(collection, query, options);
+        }
+
+        return this._fetchItemsFromServer(collection, query, options)
+          .then((response) => {
+            return getCachedQuery(collection, query)
+              .then((cachedQuery) => {
+                if (cachedQuery && response.headers) {
+                  cachedQuery.lastRequest = response.headers.requestStart;
+                  return updateCachedQuery(cachedQuery);
+                }
+
+                return null;
+              })
+              .then(() => response.data ? response.data : response);
+          })
+          .then((data) => this._replaceOfflineEntities(collection, query, data).then((data) => data.length));
+      })
+      .catch((error) => {
+        if (error instanceof InvalidCachedQuery) {
+          return getCachedQuery(collection, query)
+            .then((cachedQuery) => deleteCachedQuery(cachedQuery))
+            .catch((error) => {
+              if (error instanceof NotFoundError) {
+                return null;
+              }
+
+              throw error;
+            })
+            .then(() => this.pull(collection, query, Object.assign(options, { useDeltaSet: false })));
+        }
+
+        throw error;
+      });
   }
 
   getSyncItemCount(collection) {
@@ -150,6 +207,11 @@ export class SyncManager {
   _replaceOfflineEntities(collection, deleteOfflineQuery, networkEntities = []) {
     // TODO: this can potentially be deleteOfflineQuery.and().notIn(networkEntitiesIds)
     // but inmemory filtering with this filter seems to take too long
+    if (deleteOfflineQuery && (deleteOfflineQuery.hasSkip() || deleteOfflineQuery.hasLimit())) {
+      return this._getOfflineRepo()
+        .then((repo) => repo.update(collection, networkEntities));
+    }
+
     return this._deleteOfflineEntities(collection, deleteOfflineQuery)
       .then(() => this._getOfflineRepo())
       .then(repo => repo.update(collection, networkEntities));
@@ -290,7 +352,7 @@ export class SyncManager {
   }
 
   _fetchItemsFromServer(collection, query, options) {
-    return this._networkRepo.read(collection, query, options);
+    return this._networkRepo.read(collection, query, Object.assign(options, { dataOnly: false }));
   }
 
   _getOfflineRepo() {
@@ -365,8 +427,8 @@ export class SyncManager {
 
   _getInternalPullQuery(userQuery, totalCount) {
     userQuery = userQuery || {};
-    const { filter, sort, fields, skip } = userQuery;
-    const query = new Query({ filter, sort, fields, skip });
+    const { filter, sort, fields } = userQuery;
+    const query = new Query({ filter, sort, fields });
     query.limit = totalCount;
 
     if (!sort || isEmpty(sort)) {
@@ -396,27 +458,40 @@ export class SyncManager {
 
   _getExpectedEntityCount(collection, userQuery) {
     const countQuery = new Query({ filter: userQuery.filter });
-    return this._networkRepo.count(collection, countQuery)
-      .then(totalCount => {
-        return Math.min(totalCount - userQuery.skip, userQuery.limit || Infinity);
+    return this._networkRepo.count(collection, countQuery, { dataOnly: false })
+      .then((response) => {
+        return {
+          lastRequest: response.headers ? response.headers.requestStart : undefined,
+          count: response.data ? response.data.count : response
+        };
       });
   }
 
   _paginatedPull(collection, userQuery, options = {}) {
     let pullQuery;
-    let expectedCount;
     userQuery = userQuery || new Query();
     return this._getExpectedEntityCount(collection, userQuery)
-      .then((count) => {
-        expectedCount = count;
-        pullQuery = this._getInternalPullQuery(userQuery, expectedCount);
-        return this._deleteOfflineEntities(collection, pullQuery);
-      })
-      .then(() => {
-        const pageSizeSetting = options.autoPagination && options.autoPagination.pageSize;
-        const pageSize = pageSizeSetting || maxEntityLimit;
-        const paginatedQueries = splitQueryIntoPages(pullQuery, pageSize, expectedCount);
-        return this._executePaginationQueries(collection, paginatedQueries, options);
+      .then(({ lastRequest, count }) => {
+        pullQuery = this._getInternalPullQuery(userQuery, count);
+        return this._deleteOfflineEntities(collection, pullQuery)
+          .then(() => {
+            const pageSizeSetting = options.autoPagination && options.autoPagination.pageSize;
+            const pageSize = pageSizeSetting || maxEntityLimit;
+            const paginatedQueries = splitQueryIntoPages(pullQuery, pageSize, count);
+            return this._executePaginationQueries(collection, paginatedQueries, options);
+          })
+          .then((result) => {
+            return getCachedQuery(collection, userQuery)
+              .then((cachedQuery) => {
+                if (cachedQuery) {
+                  cachedQuery.lastRequest = lastRequest;
+                  return updateCachedQuery(cachedQuery);
+                }
+
+                return null;
+              })
+              .then(() => result);
+          });
       });
   }
 }
